minors

+++ lensit/ffs_covs/ell_mat.py
@@ -6,34 +6,28 @@ import pickle as pk
 
 import pyfftw
 
-from lensit.sims import sims_generic
 from lensit.sims.sims_generic import hash_check
-from lensit.misc.misc_utils import npy_hash
+from lensit.misc.misc_utils import npy_hash, Freq
 from lensit.misc.rfft2_utils import udgrade_rfft2, supersample
 from lensit.pbs import pbs
 
 
-def Freq(i, N):
-    """
-     Outputs the absolute integers frequencies [0,1,...,N/2,N/2-1,...,1]
-     in numpy fft convention as integer i runs from 0 to N-1.
-     Inputs can be numpy arrays e.g. i (i1,i2,i3) with N (N1,N2,N3)
-                                  or i (i1,i2,...) with N
-     Both inputs must be integers.
-     All entries of N must be even.
-    """
-    assert np.all(N % 2 == 0), "This routine only for even numbers of points"
-    return i - 2 * (i >= (N // 2)) * (i % (N // 2))
-
-
 class ell_mat:
-    """
-    Class to setup and handle Fourier mode structure on the flat sky, at the given resolution and size of
-    the rectangular box. Each 2d vector kx,ky is assigned to a multipole ell according to ell = int(|k| - 1/2).
-    Caches various matrices to facilitate multiple calls.
+    """Library helping with flat-sky patch discretization and harmonic mode structure.
+
+        This handles Fourier mode structure on the flat sky, at the given resolution and size of
+        specified rectangular box.
+
+        Args:
+            lib_dir: various things might be cached there
+                    (unlens *cache* is set to 0, in which case only a hashdict is put there at instantiation)
+            shape(2-tuple): pair of int defining the number of pixels on each side of the box
+            lsides(2-tuple): physical size (in radians) of the box sides
+            cache(optional): if non-zero, a bunch of matrices might be cached to speed up some calculations.
+
     """
 
-    def __init__(self, lib_dir, shape, lsides,mmap_mode = None):
+    def __init__(self, lib_dir, shape, lsides, cache=0):
         assert len(shape) == 2 and len(lsides) == 2
         assert shape[0] % 2 == 0 and shape[1] % 2 == 0
         assert shape[0] < 2 ** 16 and shape[1] < 2 ** 16
@@ -41,7 +35,8 @@ class ell_mat:
         self.rshape = (shape[0], shape[1] // 2 + 1)
         self.lsides = tuple(lsides)
         self.lib_dir = lib_dir
-        self.mmap_mode = mmap_mode
+        self.mmap_mode = None
+        self.cache=cache
 
         # Dumping ell mat in lib_dir. Overwrites if already present.
         fn_hash = os.path.join(lib_dir, "ellmat_hash.pk")
@@ -53,11 +48,11 @@ class ell_mat:
 
         hash_check(pk.load(open(fn_hash, 'rb')), self.hash_dict())
 
-        if pbs.rank == 0 and not os.path.exists(os.path.join(self.lib_dir, 'ellmat.npy')):
+        if pbs.rank == 0 and self.cache > 0 and not os.path.exists(os.path.join(self.lib_dir, 'ellmat.npy')):
                 print('ell_mat:caching ells in ' + os.path.join(self.lib_dir, 'ellmat.npy'))
                 np.save(os.path.join(self.lib_dir, 'ellmat.npy'), self._build_ellmat())
         pbs.barrier()
-        # FIXME
+
         self.ellmax = int(self._get_ellmax())
         self._ell_counts = self._build_ell_counts()
         self._nz_counts = self._ell_counts.nonzero()
@@ -81,19 +76,13 @@ class ell_mat:
 
     @staticmethod
     def k2ell(k):
+        r"""Mapping of 2d-frequency to multipole :math:`\ell`
+
+            :math:`\ell = \rm{int}\left(|k| - \frac 12 \right)`
+
+        """
         ret = np.uint16(np.round(k - 0.5) + 0.5 * ((k - 0.5) < 0))
         return ret
-
-    @staticmethod
-    def pbssafe_save(fname, data_to_save, pbs_rank=None):
-        if pbs_rank is not None and pbs.rank != pbs_rank:
-            return
-        np.save(fname, data_to_save)
-        return
-
-    def check_compatible(self, ellmat):
-        hash_check(self.hash_dict(), ellmat.hash_dict())
-
 
     def __call__(self, *args, **kwargs):
         return self.get_ellmat(*args, **kwargs)
@@ -102,7 +91,9 @@ class ell_mat:
         return self.get_ellmat()[item]
 
     def get_pixwinmat(self):
-        """ sin(kx Lcell_x / 2) sin (k_y Lcell_y / 2 ) """
+        r""" sin(kx Lcell_x / 2) sin (k_y Lcell_y / 2 )
+
+        """
         ky = (np.pi/self.shape[0]) * Freq(np.arange(self.shape[0]), self.shape[0])
         ky[self.shape[0] // 2:] *= -1.
         kx = (np.pi/self.shape[1]) * Freq(np.arange(self.rshape[1]), self.shape[1])
@@ -113,9 +104,13 @@ class ell_mat:
         return np.outer(rety,retx)
 
     def get_ellmat(self, ellmax=None):
+        r"""Returns the matrix containing the multipole ell assigned to 2d-frequency k = (kx,ky)
+
         """
-        Returns the matrix containing the multipole ell assigned to k = (kx,ky)
-        """
+        if self.cache == 0:
+            ret = self._build_ellmat()
+            return ret if ellmax is None else ret[np.where(ret <= ellmax)]
+
         if ellmax is None:
             return np.load(os.path.join(self.lib_dir, 'ellmat.npy'), mmap_mode=self.mmap_mode)
         else:
@@ -123,14 +118,17 @@ class ell_mat:
             if os.path.exists(fname): return np.load(fname, mmap_mode=self.mmap_mode)
             if pbs.rank == 0:
                 print('ell_mat:caching ells in ' + fname)
-                np.save(fname, self.get_ellmat()[np.where(self.get_ellmat() <= ellmax)])
+                np.save(fname, self.get_ellmat()[np.where(self() <= ellmax)])
             pbs.barrier()
             return np.load(fname, mmap_mode=self.mmap_mode)
 
     def get_phasemat(self, ellmax=None):
+        r"""Returns the matrix containing the phase k = 'k' e^i phi
+
         """
-        Returns the matrix containing the phase k = 'k' e^i phi
-        """
+        if self.cache == 0:
+            ret = np.arctan2(self.get_ky_mat(), self.get_kx_mat())
+            return ret if ellmax is None else ret[np.where(self() <= ellmax)]
         if ellmax is None:
             fname = os.path.join(self.lib_dir, 'phasemat.npy')
             if os.path.exists(fname): return np.load(fname, mmap_mode=self.mmap_mode)
@@ -148,9 +146,12 @@ class ell_mat:
             return np.load(fname, mmap_mode=self.mmap_mode)
 
     def get_e2iphi_mat(self, cache_only=False):
+        r"""Returns the matrix containing the phase e^{2 i phi}
+
         """
-        Built such that it should hit the pbs barrier only if the matrix is not already cached.
-        """
+        if self.cache == 0:
+            return np.exp(2j * np.arctan2(self.get_ky_mat(), self.get_kx_mat()))
+
         fname = os.path.join(self.lib_dir, 'e2iphimat.npy')
         if os.path.exists(fname):
             return None if cache_only else np.load(fname, mmap_mode=self.mmap_mode)
@@ -166,26 +167,34 @@ class ell_mat:
         return ell_mat(lib_dir, LDshape, self.lsides)
 
     def get_cossin_2iphi_mat(self):
+        """
+
+            Returns:
+                :math:`\cos(2i\phi), \sin(2i\phi)` as rfft arrays
+
+        """
         e2iphi = self.get_e2iphi_mat()
         return e2iphi.real, e2iphi.imag
 
 
     def _get_ellmax(self):
-        """ Max. ell present in the grid """
+        r""" Max. ell present in the grid
+
+        """
         return np.max(self.get_ellmat())
 
     def _build_ell_counts(self):
-        """
-        Number of non-redundant entries in freq map. for each ell, in the rfftmap.
-        Corresponds roughly to ell + 1/2.
-        """
         counts = np.bincount(self.get_ellmat()[:, 1:self.rshape[1] - 1].flatten(), minlength=self.ellmax + 1)
         s_counts = np.bincount(self.get_ellmat()[0:self.shape[0] // 2 + 1, [-1, 0]].flatten())
         counts[0:len(s_counts)] += s_counts
         return counts
 
     def get_ellcounts(self):
-        # This is in fact close to ell + 1/2 and not 2ell + 1.
+        r"""Number of non-redundant entries in freq map. for each ell, in the rfftmap.
+
+            Corresponds roughly to :math:`\ell + \frac 1/2`.
+
+        """
         return self._ell_counts
 
     def get_Nell(self):
@@ -199,22 +208,22 @@ class ell_mat:
         return self._nz_counts
 
     def map2cl(self, m, m2=None):
-        """Returns Cl estimates from a map.
+        r"""Returns Cl estimates from a map.
 
             Returns a cross-cl if m2 is set. Must have compatible shape.
 
         """
         assert m.shape == self.shape, m.shape
         if m2 is not None: assert m2.shape == self.shape, m2.shape
-        return self.rfft2cl(np.fft.rfft2(m)) if map2 is None else \
-            self.rfft2cl(np.fft.rfft2(m), rfftmap2=np.fft.rfft2(m2))
+        return self._rfft2cl(np.fft.rfft2(m)) if map2 is None else \
+            self._rfft2cl(np.fft.rfft2(m), rfftmap2=np.fft.rfft2(m2))
 
-    def rfft2cl(self, rfftmap, rfftmap2=None):
-        """
-        Returns Cl estimates from a rfftmap.
-        (e.g. np.fft.rfft2(sim) where sim is the output of this library)
-        Included in self for convenience.
-        Must have the same shape then self.rshape
+    def _rfft2cl(self, rfftmap, rfftmap2=None):
+        """Returns Cl estimates from a rfftmap.
+
+            (e.g. np.fft.rfft2(sim) where sim is the output of this library)
+            Must have the same shape then self.rshape
+
         """
         assert rfftmap.shape == self.rshape, rfftmap.shape
         if rfftmap2 is not None: assert rfftmap2.shape == self.rshape, rfftmap2.shape
@@ -227,9 +236,10 @@ class ell_mat:
         return Cl
 
     def bin_inell(self, rfftmap):
-        """
-        Included in self for convenience.
-        Must have the same shape then self.rshape
+        """Bin in a rfftmap according to multipole :math:`\ell`
+
+            Input must have the same shape then self.rshape
+
         """
         assert rfftmap.shape == self.rshape, rfftmap.shape
         weights = rfftmap
@@ -274,19 +284,20 @@ class ell_mat:
                          return_index=return_index, return_inverse=return_inverse, return_counts=return_counts)
 
     def rfftmap2alm(self, rfftmap, filt_func=lambda ell: True):
-        """
-        Returns a ffs_alm array from a rfftmap.
-        :param rfftmap:
-        :return:
+        """Turns a rfft map to ffs_alm array.
+        
+            This performs the filtering defined by the instance together with the normalization
+            
+
         """
         cond = filt_func(np.arange(self.ellmax + 1))
         return rfftmap[cond[self.get_ellmat()]] * np.sqrt(np.prod(self.lsides)) / np.prod(self.shape)
 
     def alm2rfftmap(self, alm, filt_func=lambda ell: True):
-        """
-        Returns a ffs_alm array from a rfftmap.
-        :param rfftmap:
-        :return:
+        """Returns a ffs_alm array from a rfftmap.
+
+            Pseudo-inverse to *fftmap2alm*
+
         """
         cond = filt_func(np.arange(self.ellmax + 1))
         ret = np.zeros(self.rshape, dtype=complex)
@@ -294,7 +305,9 @@ class ell_mat:
         return ret
 
     def rfft2_reals(self):
-        """ Pure reals modes in 2d rfft array. (from real map shape, not rfft array) """
+        """Pure reals modes in 2d rfft according to patch specifics
+
+        """
         N0, N1 = self.shape
         fx = [0]
         fy = [0]
@@ -309,23 +322,16 @@ class ell_mat:
             fy.append(N1 // 2)
         return np.array(fx), np.array(fy)
 
-    def QU2EBrfft(self, QUmap):
-        """
-        Turns QU map list into EB rfft maps.
-        """
-        assert QUmap[0].shape == self.shape and QUmap[1].shape == self.shape
-        Qk = np.fft.rfft2(QUmap[0])
-        Uk = np.fft.rfft2(QUmap[1])
-        cos, sin = self.get_cossin_2iphi_mat()
-        return cos * Qk + sin * Uk, -sin * Qk + cos * Uk
-
 
 class ffs_alm(object):
-    """
-    Simple-minded library to facilitate filtering operations on flat-sky alms in the ffs scheme.
+    """Library to facilitate filtering operations on flat-sky alms in the ffs scheme.
+
+        The instance contains info on the sky patch size and discretization.
+        Methods includes alm2map and map2alm methods amon other things.
+
     """
 
-    def __init__(self, ellmat, filt_func=lambda ell: ell > 0,kxfilt_func = None,kyfilt_func = None):
+    def __init__(self, ellmat, filt_func=lambda ell: ell > 0, kxfilt_func = None, kyfilt_func = None):
         """
         :param ellmat: ell_mat instance that defines the mode structure in the ffs scheme.
         :param filt: is callable with Boolean return. filt(ell) tells whether or not mode ell is considered
@@ -431,6 +437,7 @@ class ffs_alm(object):
             return  npy_hash(self.filt_func(np.arange(self.ellmax + 1)))
         else:
             assert 0,'implement this'
+
     def clone(self):
         return ffs_alm(self.ell_mat, filt_func=self.filt_func)
 
@@ -487,7 +494,12 @@ class ffs_alm(object):
                          return_index=return_index, return_inverse=return_inverse, return_counts=return_counts)
 
     def get_Nell(self):
-        """ Analog of 2ell + 1 on the flat sky, inclusive filtering """
+        r"""Mode number counts on the flat-sky patch.
+
+            Analog of :math:`2\ell + 1` on the flat-sky.
+            instance mode filtering is included (i.e. set to zero for mods filtered out)
+
+        """
         Nell = 2 * self._get_ell_counts()
         for ell,cond in zip(self.ell_mat()[self.ell_mat.rfft2_reals()],self._cond()[self.ell_mat.rfft2_reals()]):
             Nell[ell] -= cond
@@ -499,9 +511,10 @@ class ffs_alm(object):
         return self.__ellcounts
 
     def _build_ell_counts(self):
-        """
-        Number of non-redundant entries in freq map. for each ell, in the rfftmap.
-        Corresponds roughly to ell + 1/2.
+        r"""Number of non-redundant entries in freq map. for each ell, in the rfftmap.
+
+            Corresponds roughly to :math:`\ell + \frac 1 2`.
+
         """
         weights = self._cond()
         counts = np.bincount(self.ell_mat()[:, 1:self.ell_mat.rshape[1] - 1].flatten(), minlength=self.ell_mat.ellmax + 1,weights=weights[:, 1:self.ell_mat.rshape[1] - 1].flatten())
@@ -510,11 +523,12 @@ class ffs_alm(object):
         return counts
 
     def alm2Pk_minimal(self, alm):
-        """
-        Overkill power spectrum estimation on the grid, with minimal binning (only exact same frequencies).
-        Many bins will have only number count 4 or something.
-        Outputs a list with integer array of squared frequencies,integer array number counts, and Pk estimates.
-        (only non zero counts frequencies)
+        """Power spectrum estimation on the grid, with minimal binning (only exact same frequencies).
+
+            Many bins will have only number count 4 or something.
+            Outputs a list with integer array of squared frequencies,integer array number counts, and Pk estimates.
+            (only non zero counts frequencies)
+
         """
         almmap = self.alm2almmap(alm)
         assert len(almmap.shape) == 2
@@ -537,8 +551,10 @@ class ffs_alm(object):
         return np.sqrt(sqd_freq) * kmin, counts[sqd_freq], Pk[sqd_freq] / counts[sqd_freq]
 
     def bicubic_prefilter(self, alm):
-        """
-        Prefilter the map for use in bicubic spline prefiltering. Useful for lensing of maps.
+        """Refilter the map for use in bicubic spline prefiltering.
+
+            Mostly useful for lensing of maps.
+
         """
         assert alm.size == self.alm_size, (alm.size, self.alm_size)
         s0, s1 = self.ell_mat.shape
@@ -616,11 +632,8 @@ class ffs_alm(object):
         return Cl[:(ellmax or self.ellmax) + 1]
 
     def udgrade(self, lib_alm, alm):
-        """
-        Degrades or upgrades a alm vector from lib_alm to the self lib_alm
-        :param lib_alm:
-        :param alm:
-        :return:
+        """Degrades or upgrades a alm vector from this *ffs_alm* instance to another instance.
+
         """
         # FIXME : high time to devise a better flat sky alm scheme, this one becomes fairly convoluted.
         if self.iseq(lib_alm, allow_shape=True): return alm
@@ -629,74 +642,41 @@ class ffs_alm(object):
         return self.almmap2alm(udgrade_rfft2(lib_alm.alm2almmap(alm), self.ell_mat.shape))
 
     def QUlms2EBalms(self, QUlms):
-        """
-        Turns QU alms list into EB alms.
+        """ Turns QU alms list into EB alms.
+
         """
         assert QUlms.shape == (2, self.alm_size), QUlms.shape
         cos, sin = self.get_cossin_2iphi()
         return np.array([cos * QUlms[0] + sin * QUlms[1], -sin * QUlms[0] + cos * QUlms[1]])
 
     def TQUlms2TEBalms(self, TQUlms):
-        """
-        Turns TQU alms list into TEB alms.
+        """Turns TQU alms list into TEB alms.
+
         """
         assert TQUlms.shape == (3, self.alm_size), TQUlms.shape
         cos, sin = self.get_cossin_2iphi()
         return np.array([TQUlms[0], cos * TQUlms[1] + sin * TQUlms[2], -sin * TQUlms[1] + cos * TQUlms[2]])
 
     def EBlms2QUalms(self, EBlms):
-        """
-        Turns EB alms list into QU alms.
+        """Turns EB alms list into QU alms.
+
         """
         assert EBlms.shape == (2, self.alm_size), EBlms.shape
         cos, sin = self.get_cossin_2iphi()
         return np.array([cos * EBlms[0] - sin * EBlms[1], sin * EBlms[0] + cos * EBlms[1]])
 
     def TEBlms2TQUalms(self, TEBlms):
-        """
-        Turns EB alms list into QU alms.
+        """Turns TEB alms list into TQU alms.
+
         """
         assert TEBlms.shape == (3, self.alm_size), TEBlms.shape
         cos, sin = self.get_cossin_2iphi()
         return np.array([TEBlms[0], cos * TEBlms[1] - sin * TEBlms[2], sin * TEBlms[1] + cos * TEBlms[2]])
 
 
-class ffs_lib_phas(sims_generic.sim_lib):
-    """
-    Sim lib for Gaussian phases (Cl == 1 alm vectors for ell range set by lib_alm,
-    in the alm scheme as given by lib_alm).
-    """
-
-    def __init__(self, lib_dir, nfields, lib_alm, **kwargs):
-        self.lib_dir = lib_dir
-        self.lib_alm = lib_alm
-        self.nfields = nfields
-        super(ffs_lib_phas, self).__init__(lib_dir, **kwargs)
-
-    def _build_sim_from_rng(self, rng_state, phas_only=False):
-        np.random.set_state(rng_state)
-        alms = np.array([(np.random.standard_normal(self.lib_alm.alm_size) +
-                1j * np.random.standard_normal(self.lib_alm.alm_size)) / np.sqrt(2.) for i in range(self.nfields)])
-        if phas_only: return
-        # Reality conditions on the rfft maps
-        sla = slice(self.lib_alm.ell_mat.shape[0] // 2 + 1, self.lib_alm.ell_mat.shape[0], 1)
-        slb = slice(self.lib_alm.ell_mat.shape[0] // 2 - 1, 0, -1)
-
-        for i in range(self.nfields):
-            rfft = self.lib_alm.alm2rfft(alms[i])
-            rfft[sla, [-1, 0]] = np.conjugate(rfft[slb, [-1, 0]])
-            rfft.real[self.lib_alm.ell_mat.rfft2_reals()] *= np.sqrt(2.)
-            rfft.imag[self.lib_alm.ell_mat.rfft2_reals()] = 0.
-            alms[i] = self.lib_alm.rfftmap2alm(rfft)
-        return alms
-
-    def hashdict(self):
-        return {'nfields': self.nfields, 'lib_alm': self.lib_alm.hashdict()}
-
-
 class ffs_alm_pyFFTW(ffs_alm):
-    """
-    Same as ffs_alm but ffts are done with much faster (10 x !) threaded fftw library.
+    """Same as ffs_alm but with pyfftw fft engine  threaded fftw library (up to 10 times faster than numpy's).
+
     """
 
     def __init__(self, ellmat, filt_func=lambda ell: ell > 0, num_threads=int(os.environ.get('OMP_NUM_THREADS', 1)), flags_init=('FFTW_MEASURE',)):
